# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Liberfly",
        "version": "1.0.0"
    },
    "paths": {
        "/api/product/count": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Display a count of the resource.",
                "description": "Find products quantity",
                "operationId": "c67f43c5f4b4f77adbff97b0a8e67c0a",
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/product": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Find Products",
                "operationId": "5cc57bec35d0f2045983cc8dcd00eccd",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Data page",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "take",
                        "in": "query",
                        "description": "Register Quantity per page",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Store a newly created resource in storage.",
                "description": "Create Product",
                "operationId": "1d0b0d8724685b7b8f9d87f403bfc4df",
                "requestBody": {
                    "description": "Product data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "code",
                                    "value"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "",
                                        "example": "Nome"
                                    },
                                    "code": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "10sa2"
                                    },
                                    "value": {
                                        "type": "number",
                                        "format": "string",
                                        "example": "2.50"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/product/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Display the specified resource.",
                "description": "Find Product by Id",
                "operationId": "2768d88590e39a7ded1d0032e64dded7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product Id.",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Update the specified resource in storage.",
                "description": "Change Product",
                "operationId": "d3467c795490afbb49680821c9bd6b0d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Product data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "code",
                                    "value"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "",
                                        "example": "Nome"
                                    },
                                    "code": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "10sa2"
                                    },
                                    "value": {
                                        "type": "number",
                                        "format": "string",
                                        "example": "2.50"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Remove the specified resource from storage.",
                "description": "Delete Product",
                "operationId": "c830155516bad1f72b0081160edeb2ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Display a listing of the resource.",
                "description": "List user data",
                "operationId": "971db9bc974c0e607288cfe7ecc10096",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Data page",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "take",
                        "in": "query",
                        "description": "Register Quantity per page",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Store a newly created resource in storage.",
                "description": "Create user",
                "operationId": "089999749fbdec98922de2f74ba8d540",
                "requestBody": {
                    "description": "User data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "format": "string",
                                        "example": ""
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Display the specified resource.",
                "description": "List user data",
                "operationId": "d8f696420bea235f4916124c9ba1c507",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update the specified resource in storage.",
                "description": "Change user",
                "operationId": "5e09c26a0abb2b1bf040badb2d9c69ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "format": "string",
                                        "example": ""
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Remove the specified resource from storage.",
                "description": "Delete user",
                "operationId": "f8d7aea516ee464d0afee90f3a1a1722",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Base Crud Controller",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}